pyproject.toml
[project]
name = "xolo-pipeline"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.12",
]




[project.scripts]
xolo = "xolo.__main__:app"


[tool.uv]
package = true
